test: LikedCafeControllerTest CustomGiven 적용
